separate cache&cache_object, add to_entry method for cache_object
fix comment
