+++ mercury/src/internal/pack/waitlist.rs
@@ -1,6 +1,6 @@
 use dashmap::DashMap;
 use venus::hash::SHA1;
-use crate::internal::pack::cache::CacheObject;
+use crate::internal::pack::cache_object::CacheObject;
 
 /// Waitlist for Delta objects while the Base object is not ready.
 /// Easier and faster than Channels.
